fix(ApplicationCommandManager): limit permission methods to guilds
